feat: improve install script a lot
 * Add a lot of validations and error messages to alert the user of what went wrong
 * Truly use sed POSIX-ly
 * Remove /src and /mysql every time we redo the compose up to prevent error,
   Docker and Composer require the folders to be empty.

+++ install.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env sh
 #
-## This script is POSIX-compliant
+# This script is POSIX-compliant
 #
 # By Diego Villarreal Nevárez, @diegovillanev in my Job GitHub profile
 
@@ -13,50 +13,51 @@ else
 	docker_compose_cmd="docker-compose"
 fi
 
-## We are gonna use this var a lot, so let's make a constant
+sudo rm -rf "$PWD/src" "$PWD/mysql";
+mkdir -pv "$PWD/src" "$PWD/mysql";
+
+## We are gonna use this var a lot, so let's make it a constant
 LARAVEL_SRC_FOLDER="$PWD/src"
-DB_FOLDER="$PWD/mysql"
 
 ## If the user is root, use the compose file that uses the root php Dockerfile
 case $(id -u) \
 in
 	0)
-		find "${DB_FOLDER:?}" -mindepth 1 -delete && \
-		${docker_compose_cmd} --file docker-compose-root.yml up --detach --build app && \
-		find "${LARAVEL_SRC_FOLDER:?}" -mindepth 1 -delete && \
-		bash -c "cd ./src && ${docker_compose_cmd} run --rm composer create-project laravel/laravel ."
+		${docker_compose_cmd:?} --file docker-compose-root.yml up --detach --build app && \
+		bash -c "cd ./src && ${docker_compose_cmd:?} run --rm composer create-project laravel/laravel ."
 	;;
 	*)
-		find "${DB_FOLDER:?}" -mindepth 1 -delete && \
-		${docker_compose_cmd} --file docker-compose.yml up --detach --build app && \
-		find "${LARAVEL_SRC_FOLDER:?}" -mindepth 1 -delete && \
-		bash -c "cd ./src && ${docker_compose_cmd} run --rm composer create-project laravel/laravel ."
+		${docker_compose_cmd:?} --file docker-compose.yml up --detach --build app && \
+		bash -c "cd ./src && ${docker_compose_cmd:?} run --rm composer create-project laravel/laravel ."
 	;;
 esac
 
-## Looks like sed won't detect hidden files, so let's convert it first
-mv -v "${LARAVEL_SRC_FOLDER:?}/.env" "${LARAVEL_SRC_FOLDER:?}/env"
+if [ -f "${LARAVEL_SRC_FOLDER:?}/bootstrap/app.php" ];
+then
+	if [ -f "${LARAVEL_SRC_FOLDER:?}/.env" ];
+	then
+		## Use POSIX sed to add our environment vars to a temp file
+		sed -E -e 's|^(DB_CONNECTION=).*|\1mysql|g' \
+		       -e 's|^# ?(DB_HOST=).*|\1mysql|g' \
+		       -e 's|^# ?(DB_DATABASE=).*|\1homestead|g' \
+		       -e 's|^# ?(DB_USERNAME=).*|\1homestead|g' \
+		       -e 's|^# ?(DB_PASSWORD=).*|\1secret|g' \
+		"${LARAVEL_SRC_FOLDER:?}/.env" > "${LARAVEL_SRC_FOLDER:?}/env.tmp"
 
-sed_command=sed
-if command -v gsed > /dev/null 2>&1; then
-	sed_command=gsed
-fi
+		## Let's convert temp file to an actual envfile so the project can read its env vars
+		mv -v "${LARAVEL_SRC_FOLDER:?}/env.tmp" "${LARAVEL_SRC_FOLDER:?}/.env"
+	else
+		printf "%b" "${LARAVEL_SRC_FOLDER:?}/.env file could no be found! Exiting...\n\n" 1>&2
+		exit 1
+	fi
 
-## Use sed to edit in-place the env to add our environment vars
-${sed_command} -i -E 's|^(DB_CONNECTION=).*|\1mysql|g' "${LARAVEL_SRC_FOLDER:?}/env"
-${sed_command} -i -E 's|^# ?(DB_HOST=).*|\1mysql|g' "${LARAVEL_SRC_FOLDER:?}/env"
-${sed_command} -i -E 's|^# ?(DB_DATABASE=).*|\1homestead|g' "${LARAVEL_SRC_FOLDER:?}/env"
-${sed_command} -i -E 's|^# ?(DB_USERNAME=).*|\1homestead|g' "${LARAVEL_SRC_FOLDER:?}/env"
-${sed_command} -i -E 's|^# ?(DB_PASSWORD=).*|\1secret|g' "${LARAVEL_SRC_FOLDER:?}/env"
+	## Run the DB migrations and seeding
+	${docker_compose_cmd:?} run --rm artisan migrate --seed
 
-## Let's convert env file back to normal so the project can read its env vars
-mv -v "${LARAVEL_SRC_FOLDER:?}/env" "${LARAVEL_SRC_FOLDER:?}/.env"
-
-## Run the DB migrations and seeding
-${docker_compose_cmd} run --rm artisan migrate --seed
-
-## Let's make the modifications to read a row from the DB
-cat << EOF > "${LARAVEL_SRC_FOLDER:?}/routes/web.php"
+	## Let's make the modifications to read a row from the DB
+	if [ -f "${LARAVEL_SRC_FOLDER:?}/routes/web.php" ];
+	then
+		cat << EOF > "${LARAVEL_SRC_FOLDER:?}/routes/web.php"
 <?php
 
 use Illuminate\Support\Facades\Route;
@@ -70,11 +71,19 @@ use App\Http\Controllers\HomeController;
 Route::get('/', [HomeController::class, 'index']);
 
 EOF
+	else
+		printf "%b" "${LARAVEL_SRC_FOLDER:?}/web/routes.php" \
+		" file could no be found! Exiting...\n\n" 1>&2
+		exit 1
+	fi
 
-## A simple controller to echo the received row because it won't return nothing
-${docker_compose_cmd} run --rm artisan make:controller HomeController
 
-cat << EOF > "${LARAVEL_SRC_FOLDER:?}/app/Http/Controllers/HomeController.php"
+	## A simple controller to echo the received row because it won't return to welcome view
+	${docker_compose_cmd:?} run --rm artisan make:controller HomeController
+
+	if [ -f "${LARAVEL_SRC_FOLDER:?}/app/Http/Controllers/HomeController.php" ];
+	then
+		cat << EOF > "${LARAVEL_SRC_FOLDER:?}/app/Http/Controllers/HomeController.php"
 <?php
 
 // app/Http/Controllers/HomeController.php
@@ -98,3 +107,14 @@ class HomeController extends Controller
 }
 
 EOF
+	else
+		printf "%b" "${LARAVEL_SRC_FOLDER:?}/app/Http/Controllers/Home" \
+		"Controller.php file could no be found! Exiting...\n\n" 1>&2
+		exit 1
+	fi
+else
+	printf "%b" "${LARAVEL_SRC_FOLDER:?}/bootstrap/app." \
+	"php file could no be found! Exiting...\n\n" 1>&2
+	exit 1
+fi
+
